bugfix for volume-restore override issue
Added a check to ensure that the volume-restore script will check for
and exit if it finds a volume exists with the same name as the one it wants to make.

This was to fix an issue where you could accidentally write into an existing volume
if you were to use this on the same volume you backed up, but at a later date, you'd be
effectively mixing new & old data.